+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## v0.3.0-prerelease - May 25th, 2026
+- Covering the detection area no longer duplicates items
+- In-depth session viewing
+- Sessions can now be deleted
+- Live Log now autoscrolls
+- Added total silver count to the current session
+- Market prices now update-able
+
+## v0.2.1-prerelease - May 23rd, 2026
+- Replaced pywebview2 with PySide6
+
 ## v0.2.0-prerelease - April 6th, 2026
 - Added pause/resume feature
 - Basic session information viewer